coin/requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
jsonrpcclient==2.5.2
jsonschema==2.6.0
pkg-resources==0.0.0
pymongo==3.6.1
python-bitcoinrpc==1.0
requests==2.18.4
urllib3==1.22
